[Data] Inflate build time
